Move Row/DayCell components into MonthGrid
